made token injector and api prefix injector work, further refactoring with authServices
