+++ honestcash-v2/src/app/core/core.module.ts
@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 import { HTTP_INTERCEPTORS, HttpClient, HttpClientModule } from '@angular/common/http';
 import { RouteReuseStrategy, RouterModule } from '@angular/router';
 import { RouteReusableStrategy } from './route-reusable-strategy';
-import { AuthenticationService } from './auth/authentication.service';
-import { AuthenticationGuard } from './auth/authentication.guard';
+import { AuthenticationService } from './services/authentication.service';
+import { AuthenticationGuard } from './guards/authentication.guard';
 import { HttpService } from './http/http.service';
 import { HttpCacheService } from './http/http-cache.service';
 import { ApiPrefixInterceptor } from './http/api-prefix.interceptor';
