+++ honestcash-v2/src/app/core/http/api-prefix.interceptor.ts
@@ -3,7 +3,6 @@ import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from '@angular/c
 import { Observable } from 'rxjs';
 
 import { environment } from '../../../environments/environment';
-import { AuthenticationService } from '../auth/authentication.service';
 
 /**
  * Prefixes all requests with `environment.apiUrl`.
@@ -12,20 +11,12 @@ import { AuthenticationService } from '../auth/authentication.service';
   providedIn: 'root'
 })
 export class ApiPrefixInterceptor implements HttpInterceptor {
-  constructor(private authentificationService: AuthenticationService) {}
+  constructor() {}
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    if (this.authentificationService.credentials) {
-      if (!/^(http|https):/i.test(request.url)) {
-        request = request.clone({
-          url: environment.apiUrl + request.url
-        });
-      }
-
+    if (!/^(http|https):/i.test(request.url)) {
       request = request.clone({
-        setHeaders: {
-          'x-auth-token': `${this.authentificationService.credentials.token}`
-        }
+        url: environment.apiUrl + request.url
       });
     }
