+++ honestcash-v2/src/app/core/http/token.interceptor.ts
@@ -1,11 +1,8 @@
 import { Injectable, Injector } from '@angular/core';
 import {
-  HttpEvent, HttpInterceptor, HttpHandler, HttpRequest,
-  HttpErrorResponse
+  HttpEvent, HttpInterceptor, HttpHandler, HttpRequest
 } from '@angular/common/http';
-import { Observable, throwError } from 'rxjs';
-import { catchError } from 'rxjs/operators';
-import { Router } from '@angular/router';
+import { Observable } from 'rxjs';
 import { AuthService } from '../services/auth.service';
 
 
@@ -16,31 +13,14 @@ export class TokenInterceptor implements HttpInterceptor {
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     this.authService = this.injector.get(AuthService);
     const token: string = this.authService.getToken();
-    request = request.clone({
-      setHeaders: {
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      }
-    });
+    if (token) {
+      request = request.clone({
+        setHeaders: {
+          'x-auth-token': `${token}`,
+          'Content-Type': 'application/json'
+        }
+      });
+    }
     return next.handle(request);
-  }
-}
-
-@Injectable()
-export class ErrorInterceptor implements HttpInterceptor {
-  constructor(private router: Router) {}
-  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-
-    return next.handle(request)
-      .pipe(
-        catchError((response: any) => {
-          if (response instanceof HttpErrorResponse && response.status === 401) {
-            localStorage.removeItem('token');
-            this.router.navigateByUrl('/login');
-          }
-          console.log('response', response);
-          return throwError(response);
-        }),
-      );
   }
 }
